requirements.txt
groq>=0.9.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
rank-bm25>=0.2.2
pypdf>=4.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
streamlit>=1.36.0
click>=8.1.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
rich>=13.0.0
httpx>=0.27.0
